+++ NEWS.md
@@ -4,7 +4,9 @@
 
 - Fixed a bug causing an assertion failure when cargo-mutants was run from a
   subdirectory of a workspace. Thanks to Adam Chalmers!
- 
+
+- Generate `HttpResponse::Ok().finish()` as a mutation of an Actix `HttpResponse`.
+
 ## 23.6.0
 
 - Generate `Box::leak(Box::new(...))` as a mutation of functions returning